Set up development environment for porting to Node+TypeScript
