+++ package.json
@@ -1,17 +1,39 @@
 {
   "name": "super-tic-tac-toe",
   "version": "1.0.0",
-  "description": "Super Tic Tac Toe: a 9x9 revamp of Tic Tac Toe with added mechanics to implement more strategy" 
-  "main": "index.js",
-  "directories": {
-    "lib": "lib"
+  "description": "Super Tic Tac Toe: a 9x9 revamp of Tic Tac Toe with added mechanics to implement more strategy",
+  "main": "src/server.ts",
+  "scripts": {
+    "start": "npm run serve",
+    "build": "npm run build-sass && npm run build-ts && npm run tslint",
+    "serve": "node src/server.ts",
+    "watch-node": "nodemon dist/server.js",
+    "watch": "concurrently -k -p \"[{name}]\" -n \"Sass,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-sass\" \"npm run watch-ts\" \"npm run watch-node\"",
+    "test": "jest --forceExit --coverage --verbose",
+    "watch-test": "npm run test -- --watchAll",
+    "build-ts": "tsc",
+    "watch-ts": "tsc -w",
+    "build-sass": "node-sass src/public/stylesheets/ --output dist/public/stylesheets/",
+    "watch-sass": "node-sass -w src/public/stylesheets/ --output dist/public/stylesheets",
+    "tslint": "tslint -c tslint.json -p tsconfig.json",
+    "copy-static-assets": "ts-node copyStaticAssets.ts",
+    "debug": "npm run build && npm run watch-debug",
+    "serve-debug": "nodemon --inspect dist/server.js",
+    "watch-debug": "concurrently -k -p \"[{name}]\" -n \"Sass,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-sass\" \"npm run watch-ts\" \"npm run serve-debug\""
   },
   "dependencies": {
+    "dotenv": "^6.0.0",
+    "express": "^4.16.3",
+    "lodash": "^4.17.10",
     "svg.js": "^2.6.4"
   },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+  "devDependencies": {
+    "@types/dotenv": "^4.0.3",
+    "@types/express": "^4.16.0",
+    "node-sass": "^4.9.0",
+    "nodemon": "^1.17.5",
+    "tslint": "^5.10.0",
+    "typescript": "^2.9.2"
   },
   "repository": {
     "type": "git",
