+++ src/public/scripts/game/winCheck.js
