+++ src/public/scripts/rules.js
